% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CV.twostageSL.R
\name{plot.CV.twostageSL}
\alias{plot.CV.twostageSL}
\title{Graphical Display Of The V-Fold CV Risk Estimates}
\usage{
\method{plot}{CV.twostageSL}(x, package = "ggplot2", constant = qnorm(0.975), sort = TRUE, ...)
}
\arguments{
\item{x}{The output from \code{CV.twostageSL}.}

\item{package}{Either "ggplot2" or "lattice". The package selected must be available.}

\item{constant}{A numeric value. The confidence interval is defined as p +/- constant * se, where p is the point estimate and se is the standard error. The default is the quantile of the standard normal corresponding to a 95\% CI.}

\item{sort}{Logical. Should the rows in the plot be sorted from the smallest to the largest point estimate. If FALSE, then the order is two stage super learner, discrete super learner, then the estimators in \code{library.2stage} and \code{library.1stage} and the last is standard super learner.}

\item{...}{Additional arguments for \code{summary.CV.twostageSL}}
}
\value{
Returns the plot (either a ggplot2 object (class \code{ggplot}) or a lattice object (class \code{trellis}))
}
\description{
The function plots the V-fold cross-validated risk estimates for the two stage super learner, the standard super learner, the discrete super learner and each algorithm in the library. By default the estimates will be sorted and include an asymptotic 95% confidence interval.
}
\details{
see \code{summary.CV.twostageSL} for details on how the estimates are computed
}
\seealso{
\code{\link{summary.CV.twostageSL}} and \code{\link{CV.twostageSL}}
}
\author{
Ziyue Wu
}
