% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.coxph.R
\name{predict.SL.coxph}
\alias{predict.SL.coxph}
\title{Prediction method for cox proportional hazard regression}
\usage{
\method{predict}{SL.coxph}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{SL.coxph}}

\item{newdata}{A \code{data.frame} to generate predictions for}

\item{...}{Other arguments (unused)}
}
\value{
A numeric vector of predictions
}
\description{
Prediction method for cox proportional hazard regression
}
\examples{
# load cost data
data(cost_data)
# fit cox ph model
fit_coxph <- SL.coxph(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_coxph <- predict(fit_coxph$fit, newdata = cost_data[,c("female", "race")])
}
