% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.gilleskie.R
\name{predict.SL.gilleskie}
\alias{predict.SL.gilleskie}
\title{Prediction method for \code{SL.gilleskie}}
\usage{
\method{predict}{SL.gilleskie}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{SL.gilleskie}}

\item{newdata}{A \code{data.frame} for which predictions are desired}

\item{...}{Other arguments (unused)}
}
\description{
Prediction method for \code{SL.gilleskie}
}
\examples{
# load cost data
data(cost_data)
# fit gilleskie model
fit_gilleskie <- SL.gilleskie(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_gilleskie <- predict(fit_gilleskie$fit, newdata = cost_data[,c("female", "race")])
}
