% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.smearglm.R
\name{predict.SL.smearglm}
\alias{predict.SL.smearglm}
\title{Prediction method for \code{SL.smearglm}}
\usage{
\method{predict}{SL.smearglm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{SL.smearglm}}

\item{newdata}{A \code{data.frame} to generate predictions for}

\item{...}{Other arguments (unused)}
}
\value{
A numeric vector of predictions
}
\description{
Prediction method for \code{SL.smearglm}
}
\examples{
# load cost data
data(cost_data)
# fit smear estimator
fit_smearglm <- SL.smearglm(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_smearglm <- predict(fit_smearglm$fit, newdata = cost_data[,c("female", "race")])
}
