% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.wang.R
\name{predict.SL.wang}
\alias{predict.SL.wang}
\title{Prediction method for \code{SL.wang}}
\usage{
\method{predict}{SL.wang}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{SL.wang}}

\item{newdata}{A \code{data.frame} to generate predictions for}

\item{...}{Other arguments (unused)}
}
\value{
A numeric vector of predictions
}
\description{
Prediction method for \code{SL.wang}
}
\examples{
# load cost data
data(cost_data)
# fit wang model
fit_wang <- SL.wang(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_wang <- predict(fit_wang$fit, newdata = cost_data[,c("female", "race")])
}
