% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.coxph.R
\name{SL.coxph}
\alias{SL.coxph}
\title{Cox proportional hazard regression models back-transformed to the conditional mean}
\usage{
SL.coxph(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights (not currently used)}

\item{...}{Other arguments (unused)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted \code{coxph} model object),
                   \code{randomUpper}.}
}
}
\description{
This function implements Cox proportional hazard regression models (David 1972),
and back-transforms to the conditional mean scale (see also Basu 2005).
}
\examples{
# load cost data
data(cost_data)
# fit cox ph model
fit_coxph <- SL.coxph(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_coxph <- predict(fit_coxph$fit, newdata = cost_data[,c("female", "race")])
}
\references{
Basu A and Rathouz PJ. Estimating marginal and incremental effects on health outcomes using flexible link and variance function models. Biostatistics 2005; 6(1): 93–109.
}
