% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.flexsurv.R
\name{SL.flexsurv.gengamma}
\alias{SL.flexsurv.gengamma}
\title{Accelerated failure time model with generalized gamma family.}
\usage{
SL.flexsurv.gengamma(..., dist = "gengamma")
}
\arguments{
\item{...}{Other arguments passed to \code{\link[twostageSL]{SL.flexsurv}}.}

\item{link}{Link function}
}
\description{
Accelerated failure time model with generalized gamma family.
}
\references{
Manning WG, Basu A and Mullahy J. Generalized modeling approaches to risk adjustment of skewed outcomes data. Journal of Health Economics 2005; 24(3): 465–488.
}
