% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.flexsurv.R
\name{SL.flexsurv.lognormal}
\alias{SL.flexsurv.lognormal}
\title{Accelerated failure time model with log-Normal family.}
\usage{
SL.flexsurv.lognormal(..., dist = "lognormal")
}
\arguments{
\item{...}{Other arguments passed to \code{SL.flexsurv}.}

\item{link}{Link function}
}
\description{
See Manning (2005) and \code{\link[twostageSL]{SL.flexsurv}}.
}
\examples{
# load cost data
data(cost_data)
# fit AFT model with weibull family
fit_flexsurv.lognormal <- SL.flexsurv.lognormal(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_flexsurv.lognormal <- predict(fit_flexsurv.lognormal$fit, newdata = cost_data[,c("female", "race")])
}
\references{
Manning WG, Basu A and Mullahy J. Generalized modeling approaches to risk adjustment of skewed outcomes data. Journal of Health Economics 2005; 24(3): 465–488.
}
