% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.flexsurv.R
\name{SL.flexsurv.weibull}
\alias{SL.flexsurv.weibull}
\title{Accelerated failure time model with Weibull family.}
\usage{
SL.flexsurv.weibull(..., dist = "weibull")
}
\arguments{
\item{...}{Other arguments passed to \code{\link[twostageSL]{SL.flexsurv}}.}

\item{link}{Link function}
}
\description{
Accelerated failure time model with Weibull family.
}
\examples{
# load cost data
data(cost_data)
# fit AFT model with weibull family
fit_flexsurv.weibull <- SL.flexsurv.weibull(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_flexsurv.weibull <- predict(fit_flexsurv.weibull$fit, newdata = cost_data[,c("female", "race")])
}
\references{
Manning WG, Basu A and Mullahy J. Generalized modeling approaches to risk adjustment of skewed outcomes data. Journal of Health Economics 2005; 24(3): 465–488.
}
