% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.gilleskie.R
\name{SL.gilleskie}
\alias{SL.gilleskie}
\title{Adaptive Hazard method of Gilleskie and Mroz}
\usage{
SL.gilleskie(
  Y,
  X,
  newX,
  family = gaussian(),
  obsWeights = rep(1, length(Y)),
  kValues = c(5, 15, 25),
  maxPoly = 2,
  pValThresh = 0.05,
  ...
)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights (not currently used)}

\item{kValues}{Number of intervals to bin the variable into in order to estimate the discrete
hazard function}

\item{maxPoly}{The largest degree polynomial to be used in fitting the discrete hazard
function}

\item{pValThresh}{The threshold for p-values used in the covariate selection process}

\item{...}{Other arguments (not currently used)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted hazard regression object),
                   \code{maxK} (the selected number of partitions), and \code{hK}
                   (the mean outcome in each partition)}
}
}
\description{
This function implements the estimator developed by Gilleskie and Mroz,
which estimates the conditional mean costs by adaptively modeling the conditional hazard
function and back-transforming into an estimate of the conditional mean. For more description,
we refer users to the original paper.
}
\examples{
# load cost data
data(cost_data)

fit_gilleskie <- SL.gilleskie(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
}
\references{
Gilleskie, Donna and Mroz, Thomas, (2004), A flexible approach for estimating the effects of covariates on health expenditures, Journal of Health Economics, 23, issue 2, p. 391-418.
}
