% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.glm.gamma.R
\name{SL.glm.gamma}
\alias{SL.glm.gamma}
\title{Generalized linear model with gamma family and user-specified link function}
\usage{
SL.glm.gamma(Y, X, newX, family, obsWeights, link = "log", ...)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights (not currently used)}

\item{link}{Link function}

\item{...}{Other arguments (not currently used)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted \code{glm} regression object)}
}
}
\description{
Generalized linear model with gamma family and user-specified link function
}
\examples{
# load cost data
data(cost_data)
# fit gamma glm with log link model
fit_glm.gamma <- SL.glm.gamma(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_glm.gamma <- predict(fit_glm.gamma$fit, newdata = cost_data[,c("female", "race")])

}
