% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.glm.gamma.R
\name{SL.glm.gammaid}
\alias{SL.glm.gammaid}
\title{Generalized linear model with gamma family and identity link function}
\usage{
SL.glm.gammaid(..., link = "identity")
}
\arguments{
\item{...}{Other arguments passed to \code{SL.glm.gamma}.}

\item{link}{Link function}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted \code{glm} regression object)}
}
}
\description{
See \code{\link[twostageSL]{SL.glm.gamma}}.
}
\examples{
# load cost data
data(cost_data)
# fit gamma glm with identity link
fit_glm.gammaid <- SL.glm.gammaid(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_glm.gammaid <- predict(fit_glm.gammaid$fit, newdata = cost_data[,c("female", "race")])

}
