% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.glm.manning.R
\name{SL.glm.manning}
\alias{SL.glm.manning}
\title{Adaptive generalized linear model of Manning (2001)}
\usage{
SL.glm.manning(
  Y,
  X,
  newX,
  family = gaussian(),
  obsWeights = rep(1, length(Y)),
  kCut = 3,
  lambdaCut = c(0.5, 1.5, 2.5),
  startNLS = 0,
  ...
)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights (not currently used)}

\item{kCut}{Cut point for kurtosis}

\item{lambdaCut}{Cut points for skewness}

\item{startNLS}{Starting values for the non-linear least squares}

\item{...}{Other arguments (not currently used)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted regression model object)}
}
}
\description{
This function implements the estimator of Manning and Mullahy (2001).
The estimator adaptively selects a link function and family based on the skewness and
kurtosis of the data. Some of the family/link functions recommended by Manning are
numerically unstable and so this function returns Duan's smearing estimate if the GLM
fitting step breaks down.
}
\examples{
# load cost data
data(cost_data)
# fit manning model
fit_glm.manning <- SL.glm.manning(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_glm.manning <- predict(fit_glm.manning$fit, newdata = cost_data[,c("female", "race")])
}
\references{
Manning WG, Mullahy J (2001). “Estimating log models: to transform or not to transform?” Journal of Health Economics, 20(4), 461–494.
}
