% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.smearglm.R
\name{SL.smearglm}
\alias{SL.smearglm}
\title{Generalized linear model with Gaussian family and log link - Smearing Estimate}
\usage{
SL.smearglm(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights}

\item{...}{Other arguments (not currently used)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted \code{glm} regression object)
                   and \code{smear_factor} (the estimated smearing factor)}
}
}
\description{
Implements the "smearing" estimator of Duan (1983), which
fits a linear model on the log transformed outcome and transforms to an estimate of the
conditional mean on the original scale.
}
\examples{
# load cost data
data(cost_data)
# fit smear estimator
fit_smearglm <- SL.smearglm(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                              newX = cost_data[, c("female", "race")])
# get back predictions
pred_smearglm <- predict(fit_smearglm$fit, newdata = cost_data[,c("female", "race")])
}
\references{
Duan, N. (1983) Smearing estimate: a nonparametric retransformation method. Journal of the American Statistical Association, 78, 605-610.
}
