% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.wang.R
\name{SL.wang}
\alias{SL.wang}
\title{Quantile regression method of Wang and Zhou (2009)}
\usage{
SL.wang(
  Y,
  X,
  newX,
  family = gaussian(),
  obsWeights = rep(1, length(Y)),
  g = "log",
  m = length(Y),
  c = 0.2,
  b = 0.05,
  ...
)
}
\arguments{
\item{Y}{A numeric outcome variable}

\item{X}{A \code{data.frame} of covariates constituting the training sample}

\item{newX}{A \code{data.frame} with the same column names and format as \code{X} constituting
the validation sample.}

\item{family}{Gaussian only}

\item{obsWeights}{Observation-level weights (not currently used)}

\item{g}{Transformation to apply to \code{Y} before quantile regression is used. Choices
are \code{"log"} or \code{"sqrt"}}

\item{m}{Number of quantiles to compute}

\item{c}{Constant used to determine truncation level for transforming quantiles to conditional
mean}

\item{b}{Constant used to determine truncation level for transforming quantiles to conditional
mean}

\item{...}{Other arguments (not currently used)}
}
\value{
\describe{
 \item{\code{pred}}{Predicted outcomes based on predictors in \code{newX}}
 \item{\code{fit}}{A list with named entries \code{object} (the fitted \code{rq} regression object),
                   \code{alpha} (the controlled level of trimming based), and \code{g_inv}
                   (the inverse function of the inputted \code{g})}
}
}
\description{
This function implements the estimator developed by Wang and Zhou.
The estimator estimates the conditional mean costs by modeling the conditional quantiles of a transformed cost.
Using the equivariance property of quantiles to monotone transformations, the quantile estimators may be
transformed back to the conditional mean cost on the original scale.
}
\examples{
# load cost data
data(cost_data)
# fit wang model
fit_wang <- SL.wang(Y = cost_data$totalcost, X = cost_data[, c("female", "race")],
                    newX = cost_data[, c("female", "race")])
}
\references{
Wang HJ and Zhou XH. Estimation of the retransformed conditional mean in health care cost studies. Biometrika 2010; 97(1): 147–158.
}
