% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{twostageSL.control}
\alias{twostageSL.control}
\title{Control parameters for the twostageSL}
\usage{
twostageSL.control(
  saveFitLibrary = TRUE,
  saveCVFitLibrary = TRUE,
  trimLogit = 0.001
)
}
\arguments{
\item{saveFitLibrary}{Logical. Should the fit for each algorithm be saved in the output from \code{twostageSL}.}

\item{saveCVFitLibrary}{Logical. Should cross-validated fits for each algorithm be saved in the output from \code{twostageSL}.}

\item{trimLogit}{number between 0.0 and 0.5. What level to truncate the logit transformation to maintain a bounded loss function when using the NNloglik method.}
}
\value{
A list containing the control parameters.
}
\description{
Control parameters for the twostageSL
}
